requirements.txt
## These requirements were autogenerated by pipenv# To regenerate from the project's Pipfile, run:##    pipenv lock --requirements#-i https://pypi.org/simplearrow==0.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'binaryornot==0.4.4botocore==1.19.11certifi==2020.6.20chardet==3.0.4click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'configparser==5.0.1cookiecutter==1.7.2dnspython==2.0.0grpc-interceptor==0.12.0grpcio-health-checking==1.34.0grpcio-tools==1.34.0grpcio==1.34.0idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'jinja2-time==0.2.0jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'jmespath==0.10.0markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'marshmallow==3.9.1minio==6.0.0mongoengine==0.20.0poyo==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'protobuf==3.13.0pymongo==3.11.0python-dateutil==2.8.1python-decouple==3.3python-slugify==4.0.1pytz==2020.4requests==2.25.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'six==1.15.0text-unidecode==1.3urllib3==1.25.11